Move gateway control protocol into `gateway.rs`
