+++ rust/connlib/tunnel/src/gateway.rs
@@ -1,22 +1,42 @@
-use std::task::{ready, Context, Poll};
-use std::time::Duration;
-
-use crate::control_protocol::gateway::ResourceDescription;
 use crate::device_channel::Device;
 use crate::ip_packet::MutableIpPacket;
-use crate::peer::PacketTransformGateway;
+use crate::peer::{PacketTransformGateway, Peer};
 use crate::peer_store::PeerStore;
+use crate::utils::{stun, turn};
 use crate::Tunnel;
+use boringtun::x25519::PublicKey;
 use chrono::{DateTime, Utc};
 use connlib_shared::messages::{
-    ClientId, DomainResponse, Interface as InterfaceConfig, ResourceId,
+    Answer, ClientId, ConnectionAccepted, DomainResponse, Interface as InterfaceConfig, Key, Offer,
+    Relay, ResourceId,
 };
-use connlib_shared::{Callbacks, Dname};
+use connlib_shared::{Callbacks, Dname, Error, Result};
+use ip_network::IpNetwork;
+use secrecy::{ExposeSecret as _, Secret};
 use snownet::Server;
+use std::task::{ready, Context, Poll};
+use std::time::Duration;
 use tokio::time::{interval, Interval, MissedTickBehavior};
 
 const PEERS_IPV4: &str = "100.64.0.0/11";
 const PEERS_IPV6: &str = "fd00:2021:1111::/107";
+
+/// Description of a resource that maps to a DNS record which had its domain already resolved.
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct ResolvedResourceDescriptionDns {
+    pub id: ResourceId,
+    /// Internal resource's domain name.
+    pub domain: String,
+    /// Name of the resource.
+    ///
+    /// Used only for display.
+    pub name: String,
+
+    pub addresses: Vec<IpNetwork>,
+}
+
+pub type ResourceDescription =
+    connlib_shared::messages::ResourceDescription<ResolvedResourceDescriptionDns>;
 
 impl<CB> Tunnel<CB, GatewayState, Server, ClientId>
 where
@@ -40,6 +60,84 @@ where
         tracing::debug!(ip4 = %config.ipv4, ip6 = %config.ipv6, %name, "TUN device initialized");
 
         Ok(())
+    }
+
+    /// Accept a connection request from a client.
+    ///
+    /// Sets a connection to a remote SDP, creates the local SDP
+    /// and returns it.
+    ///
+    /// # Returns
+    /// The connection details
+    #[allow(clippy::too_many_arguments)]
+    pub fn set_peer_connection_request(
+        &mut self,
+        client: ClientId,
+        key: Secret<Key>,
+        offer: Offer,
+        gateway: PublicKey,
+        ips: Vec<IpNetwork>,
+        relays: Vec<Relay>,
+        domain: Option<Dname>,
+        expires_at: Option<DateTime<Utc>>,
+        resource: ResourceDescription,
+    ) -> Result<ConnectionAccepted> {
+        let resource_addresses = match &resource {
+            ResourceDescription::Dns(r) => {
+                let Some(domain) = domain.clone() else {
+                    return Err(Error::ControlProtocolError);
+                };
+
+                if !crate::dns::is_subdomain(&domain, &r.domain) {
+                    return Err(Error::InvalidResource);
+                }
+
+                r.addresses.clone()
+            }
+            ResourceDescription::Cidr(ref cidr) => vec![cidr.address],
+        };
+
+        let answer = self.connections_state.node.accept_connection(
+            client,
+            snownet::Offer {
+                session_key: key.expose_secret().0.into(),
+                credentials: snownet::Credentials {
+                    username: offer.username,
+                    password: offer.password,
+                },
+            },
+            gateway,
+            stun(&relays, |addr| {
+                self.connections_state.sockets.can_handle(addr)
+            }),
+            turn(&relays, |addr| {
+                self.connections_state.sockets.can_handle(addr)
+            }),
+        );
+
+        self.new_peer(
+            ips,
+            client,
+            resource,
+            expires_at,
+            resource_addresses.clone(),
+        )?;
+
+        tracing::info!(%client, gateway = %hex::encode(gateway.as_bytes()), "Connection is ready");
+
+        Ok(ConnectionAccepted {
+            ice_parameters: Answer {
+                username: answer.credentials.username,
+                password: answer.credentials.password,
+            },
+            domain_response: domain.map(|domain| DomainResponse {
+                domain,
+                address: resource_addresses
+                    .into_iter()
+                    .map(|ip| ip.network_address())
+                    .collect(),
+            }),
+        })
     }
 
     /// Clean up a connection to a resource.
@@ -73,6 +171,28 @@ where
         self.connections_state
             .node
             .add_remote_candidate(conn_id, ice_candidate);
+    }
+
+    fn new_peer(
+        &mut self,
+        ips: Vec<IpNetwork>,
+        client_id: ClientId,
+        resource: ResourceDescription,
+        expires_at: Option<DateTime<Utc>>,
+        resource_addresses: Vec<IpNetwork>,
+    ) -> Result<()> {
+        tracing::trace!(?ips, "new_data_channel_open");
+
+        let mut peer = Peer::new(client_id, PacketTransformGateway::default(), &ips, ());
+
+        for address in resource_addresses {
+            peer.transform
+                .add_resource(address, resource.clone(), expires_at);
+        }
+
+        self.role_state.peers.insert(peer, &ips);
+
+        Ok(())
     }
 }
